Added awscli as a requirement
I can't use boto directly since it doesn't support
Python 3. Using the new boto3 library was going to be
too much trouble and probably somewhat unstable.

So I decided to use awscli. It'll be weird, but I won't
need to use s3 too often so it shouldn't be a problem.