Added Full Report link to the bottom of the report email
Gmail sometimes fails to load one or more of the graph
images although the image is in s3 as it should be.

So I'm sending the full html of the email as well and linking
to it inside the email. That way, even if there's a problem
with images inside Gmail, I'll be able to see the full report.