Dates when no weight is entered are being filled with the last entered weight
I first implemented this app without this fill functionality.
But when I do it like that, the weight graphs move much more
drastically than how they actually should be. So I gave in
and I'm now filling the empty dates with the last measurement
entered before that date a la MyFitnessPal. The graphs look
much better that way.